Review comment. Iter 1
