+++ README.md
@@ -45,123 +45,8 @@ docker build -t rgb-lightning-node .
 
 ## UniFFI Bindings
 
-REST is **not** part of the SDK surface. REST endpoints remain a compatibility
-wrapper for existing integrations and tests. The SDK surface is exposed through
-the Rust library + UniFFI bindings.
-
-Current lifecycle/threading model:
-- UniFFI SDK is instance-based via `SdkNodeV1` object handles.
-- Create a node with `SdkNodeV1.create(SdkInitRequestV1)`, then call methods on that instance.
-- Multiple SDK node instances are supported per process.
-- Legacy global `sdk_*` wrappers remain in Rust for compatibility but are no longer part of generated UniFFI bindings.
-
-Build with UniFFI enabled:
-```sh
-cargo check --features uniffi
-```
-
-Generate all bindings:
-```sh
-./scripts/ci/uniffi_generate_all.sh
-```
-
-Generate per language:
-```sh
-./scripts/ci/uniffi_generate_kotlin.sh
-./scripts/ci/uniffi_generate_kotlin_android.sh
-./scripts/ci/uniffi_generate_python.sh
-./scripts/ci/uniffi_generate_swift.sh
-```
-
-Quick SDK smoke tests:
-```sh
-cargo test --features uniffi --lib uniffi_smoke_tests:: -- --test-threads=1
-```
-
-Library-only test suite (no daemon/API routing):
-```sh
-cargo test --features "uniffi,test-utils" \
-  --test lib_core_uniffi_state \
-  --test lib_core_type_converters \
-  --test lib_core_error_mapping \
-  --test lib_core_node_handle_invariants
-```
-
-Strict lint after library-core changes:
-```sh
-cargo clippy --all-targets --all-features -- -D warnings
-```
-
-PR CI gate for these tests:
-- `.github/workflows/uniffi-artifacts.yaml` job: `library-core-tests`
-
-Kotlin/JVM artifact build (Linux host):
-```sh
-cargo build --release --features uniffi --lib
-./scripts/ci/uniffi_generate_kotlin.sh
-```
-
-Kotlin/Android artifact build (requires `ANDROID_NDK_HOME` + `cargo-ndk`):
-```sh
-./scripts/ci/uniffi_generate_kotlin_android.sh
-./scripts/ci/build_android_jni.sh
-```
-
-Local Android NDK setup (Ubuntu, CLI):
-```sh
-# 1) Install Java + Rust helper
-sudo apt-get update
-sudo apt-get install -y openjdk-17-jdk unzip cmake clang pkg-config build-essential ninja-build
-cargo install cargo-ndk --version 3.5.4 --locked
-cargo install bindgen-cli --version 0.71.1 --locked
-
-# 2) Install Android cmdline-tools (pick your own SDK root if needed)
-export ANDROID_SDK_ROOT="$HOME/Android/Sdk"
-mkdir -p "$ANDROID_SDK_ROOT/cmdline-tools"
-cd /tmp
-wget https://dl.google.com/android/repository/commandlinetools-linux-11076708_latest.zip -O cmdline-tools.zip
-unzip -q cmdline-tools.zip
-mkdir -p "$ANDROID_SDK_ROOT/cmdline-tools/latest"
-mv cmdline-tools/* "$ANDROID_SDK_ROOT/cmdline-tools/latest/"
-
-# 3) Install NDK + build tools used by local checks/CI
-"$ANDROID_SDK_ROOT/cmdline-tools/latest/bin/sdkmanager" --licenses
-"$ANDROID_SDK_ROOT/cmdline-tools/latest/bin/sdkmanager" \
-  "ndk;26.3.11579264" "platform-tools" "build-tools;34.0.0"
-
-# 4) Export env vars for this shell
-export ANDROID_NDK_HOME="$ANDROID_SDK_ROOT/ndk/26.3.11579264"
-export PATH="$ANDROID_SDK_ROOT/cmdline-tools/latest/bin:$ANDROID_SDK_ROOT/platform-tools:$PATH"
-```
-
-Then build Android artifacts:
-```sh
-./scripts/ci/uniffi_generate_kotlin_android.sh
-./scripts/ci/build_android_jni.sh
-```
-
-Swift artifact build (macOS):
-```sh
-rustup target add aarch64-apple-ios aarch64-apple-ios-sim x86_64-apple-ios
-cargo build --release --features uniffi --lib --target aarch64-apple-ios
-cargo build --release --features uniffi --lib --target aarch64-apple-ios-sim
-cargo build --release --features uniffi --lib --target x86_64-apple-ios
-./scripts/ci/uniffi_generate_swift.sh
-./scripts/ci/package_swift_xcframework.sh
-```
-
-Parity tests used in CI:
-```sh
-cargo test zero_amount_invoice -- --test-threads=1
-cargo test send_receive -- --test-threads=1
-```
-
-CI artifact packaging workflow:
-- `.github/workflows/uniffi-artifacts.yaml`
-- Artifacts produced:
-  - Swift: `RGBLightningNode.xcframework`
-  - Kotlin JVM host bundle: generated Kotlin sources + Linux `librgb_lightning_node.so`
-  - Kotlin Android: `jniLibs` + generated Kotlin sources
+See [UniFFI SDK documentation](src/uniffi_api/README.md) for setup, runtime
+model, binding generation, test commands, and artifact packaging details.
 
 ## Run
